pyproject.toml
[project]
name = "active-planning-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "groq>=0.4.1,<1",
    "langchain>=1.2.10",
    "langchain-groq==1.1.2",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.8",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.2",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
